updated pci irq support to properly route intr_pins through the southbridge. PCI devices now have to raise irqs via the pci device, which calls a registered bridge handler...
The southbridge pci interconnect is a little obtuse. You configure the southbridge with the irq numbers for each of the 4 pci interrupt lines. When a pci device raises an irq, it raises its configured pci interrupt pin, which goes through the pci device to the southbridge, which translates that to the actual irq. So basically the southbridge sits in between the PCI bus and the CPU, which is odd because you configure the southbridge via PCI accesses.