repurpose v3_ctrl.c as a utility library for palacios userspace progs
I've taken some common idioms used in a lot of the userspace utilities
and put them in a (for now, static) library that all programs can utilize.
One example: when running an ioctl, it's no longer necessary to open the vm device,
close it, etc. This is taken care of by v3_vm_ioctl.
This should give us some more code sharing and make the userspace utilities quicker to write and a bit smaller.
For now, I've modified v3_free, v3_launch, v3_stop, and v3_create to use this library,
so you can look to those for examples.