significant refactoring of RTM emulation code
This is the first cut at a refactor of the transactional memory
extension. Most changes have been to reduce large function sizes, and
to reduce footprint in changes to Palacios core files. There are
several bug fixes as well, and this gets us to the point where we
can run thousand-transaction benchmarks on single-core.

At this point our main limitations are yet-to-be implemented
architectural features (coming soon) and limitations in the
quix86 decoder.

Signed-off-by: Maciek Swiech <dotpyfe@u.northwestern.edu>