Convert shadow paging to use 32 PAE (Core)
This changes shadow paging to use, at minimum, 32PAE
shadow page tables.   This makes it possible to place
a guest using shadow paging anywhere in host physical
memory instead of just the first 4 GB

This first patch adds core functionality to the VTLB
pager, as well as change to CR4 handling.   The
other patches with this name need to be applied for
this to work.