Convert shadow paging to use 32 PAE (Direct Paging)
This changes shadow paging to use, at minimum, 32PAE
shadow page tables.   This makes it possible to place
a guest using shadow paging anywhere in host physical
memory instead of just the first 4 GB

This patch converts the direct pager to use 32PAE.
The direct pager is active when the guest has no
installed page tables.   All patches of this name
need to be applied for this to work.