Convert shadow paging to use 32 PAE (SVM)
This changes shadow paging to use, at minimum, 32PAE
shadow page tables.   This makes it possible to place
a guest using shadow paging anywhere in host physical
memory instead of just the first 4 GB

This patch updates SVM so that the shadow pts are activated
and CR4 reads as well as writes are intercepted.

Note that all patches with this name need to be used together
for this to work.