Add swapping and pinning capability to Palacios
This commit adds allows swapping of VMs at the granularity of
base regions.  Since base regions down to the page granularity are
possible, this swapper can swap down to page granularity as well.

The functionality can be selected at compile time, and the use of
swapping is determined at run time with a new "swapping" block, like
this:

  <mem ... >N_MB</mem>             Size of memory in the GPA

  <swapping enable="y">
     <allocated>M_MB</allocated>   Allocated space (M_MB <= N_MB)
     <file>FILENAME</file>         Where to swap to
     <strategy>STRATEGY</strategy> Victim picker to use NEXT_FIT, RANDOM (default), LRU, DEFAULT
  </swapping>

Pinning is also supported