Implementation of resource control host os interface for Linux
This associates a pointer to a resource control structure with
each thread created by Palacios.   If this structure exists, its
contents are then used to control resource allocations.  Currently,
the controls are on page allocation: alignment, numa node, and
page allocation filtering.  The latter is what cache partitioning
builds on.

This also changes the /proc/v3vee/v3-guest-info-details output
slightly.  For a guest with too many cores or memory regions to
display, the output is now truncated in a graceful way.